refactor : Result의 findLastWinner메서드를 Cars 클래스로 위치 변경
